# swagger/swagger.yml
...
openapi: '3.0.3'
info:
  version: '1.0.0'
  title: Cloud service - OpenAPI 3.0
  description: |-
    This is application for a cloud service. This service is a file storage \
    in clode.
    1. User registration.\
    2. Save nad then uses user's files.
  contact:
    name: Denis
    email: work80@mail.ru
    url: https://github.com/Tryd0g0lik
servers:
  - url: http://localhost:8000/api/v1/
    description: "Basis server"
paths:
...
  /files/:
    get:
      tags:
        - file
      summary: For gat a file's list.
      description: |-
        Supeuser can get ф list from everyone of files. (Propoerty 'is_staff'
         must have the 'True' value.)
        USer caт get ф list all files fron his collection.
      operationId: FileStorageViewSet
      parameters:
        - name: user_session_<user_id>
          description: |-
            This. the user receives from registration.
          in: cookie
          required: true
          schema:
            type: string
            example: "gAWVQAAAAAAAAACMPCQyYiQxMiRNZE5uT2s2bEROSjFRd2JWeXp5dDN1NkpReTg1eDRpLlVHVUJiTGpoZTM2WkxJZGhscXloT5Qu"
        - name: is_superuser_<user_id>
          description: |-
            This. the user receives from registration.
          in: cookie
          required: true
          schema:
            type: boolean
            example: false
      responses:
        200:
            description: If all the very well
            headers:
                Set-Cookie:
                    schema:
                        allOf:
                            - type: string
                              example: user_session_<user_id>=gAWVQAAAAAAAAACMPCQyYiQxMiRNZE5uT2s2bEROSjFRd2JWeXp5dDN1NkpReTg1eDRpLlVHVUJiTGpoZTM2WkxJZGhscXloT5Qu; Path=/; HttpOnly
                              description: Cookie for user session
                            - type: string
                              example: is_superuser_<user_id>=false; Path=/; HttpOnly
                              description: Cookie for user session
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/FileProperties/oneOf/1'
                    example:
                        original_name: "README.docx"
                        size: 27981
                        upload_date: "2025-01-18T17:18:03.154301+07:00"
                        last_downloaded: null
                        comment: null
                        file_path: "/uploads/2025/1/18/README.docx"
                        special_link: "dda8958e-1711-464e-9d47-acd292241149"
        400:
          description: If not all very well
          headers:
            Set-Cookie:
              schema:
                allOf:
                  - type: string
                    example: user_session_<user_id>=gAWVQAAAAAAAAACMPCQyYiQxMiRNZE5uT2s2bEROSjFRd2JWeXp5dDN1NkpReTg1eDRpLlVHVUJiTGpoZTM2WkxJZGhscXloT5Qu; Path=/; HttpOnly
                    description: Cookie for user session
                  - type: string
                    example: is_superuser_<user_id>=false; Path=/; HttpOnly
                    description: Cookie for user session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorFile'
              example:
                error: "Mistake => < text from error >"
    post:
      tags:
        - file
      summary: Store of files
      operationId: FileStorageViewSet
      requestBody:
        required: true
        content:
          multipart/from-data:
            schema:
              $ref: '#/components/shemas/FileOfForm'
      parameters:
        - name: user_session_<user_id>
          description: |-
            This. the user receives from registration.
          in: cookie
          required: true
          schema:
            type: string
            example: "gAWVQAAAAAAAAACMPCQyYiQxMiRNZE5uT2s2bEROSjFRd2JWeXp5dDN1NkpReTg1eDRpLlVHVUJiTGpoZTM2WkxJZGhscXloT5Qu"
        - name: is_superuser_<user_id>
          description: |-
            This. the user receives from registration.
          in: cookie
          required: true
          schema:
            type: boolean
            example: false
      responses:
        200:
          description: |-
            If is all very well
          headers:
            Set-Cookies:
                schema:
                    allOf:
                        - type: string
                          example: user_session_<user_id>=gAWVQAAAAAAAAACMPCQyYiQxMiRNZE5uT2s2bEROSjFRd2JWeXp5dDN1NkpReTg1eDRpLlVHVUJiTGpoZTM2WkxJZGhscXloT5Qu; Path=/; HttpOnly
                          description: Cookie for user session
                        - type: string
                          example: is_superuser_<user_id>=false; Path=/; HttpOnly
                          description: Cookie for user session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileProperties/oneOf/0'
              example:
                id: 30
                original_name: "README.docx"
                size: 27981
                upload_date: "2025-01-18T17:18:03.154301+07:00"
                last_downloaded: null
                comment: null
                file_path: "/uploads/2025/1/18/README.docx"
                special_link: "dda8958e-1711-464e-9d47-acd292241149"
        400:
          description: |-
            If is not very well
          headers:
            Set-Cookies:
                schema:
                    allOf:
                        - type: string
                          example: user_session_<user_id>=gAWVQAAAAAAAAACMPCQyYiQxMiRNZE5uT2s2bEROSjFRd2JWeXp5dDN1NkpReTg1eDRpLlVHVUJiTGpoZTM2WkxJZGhscXloT5Qu; Path=/; HttpOnly
                          description: Cookie for user session
                        - type: string
                          example: is_superuser_<user_id>=false; Path=/; HttpOnly
                          description: Cookie for user session
          content:
            application/json:
              schema:
                properties:
                    error:
                        $ref: '#/components/schemas/ErrorFile'
              example:
                error: No file provided
        404:
          description: |-
            If is not very well
          headers:
            Set-Cookies:
                schema:
                    allOf:
                        - type: string
                          example: user_session_<user_id>=gAWVQAAAAAAAAACMPCQyYiQxMiRNZE5uT2s2bEROSjFRd2JWeXp5dDN1NkpReTg1eDRpLlVHVUJiTGpoZTM2WkxJZGhscXloT5Qu; Path=/; HttpOnly
                          description: Cookie for user session
                        - type: string
                          example: is_superuser_<user_id>=false; Path=/; HttpOnly
                          description: Cookie for user session
          content:
            application/json:
              schema:
                properties:
                    error:
                        $ref: '#/components/schemas/ErrorFile'
              example:
                error: File not found.
  /files/{pk}:
    get:
      tags:
        - file
      summary: The method recieve a single posiyion by 'pk'
      description: |-
        User can get a single position by an id  of position if hi is the owner (this id from his a collaction).
        Superuser can get a one position of everyone user.
      operationId: FileStorageViewSet
      parameters:
        - name: pk
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UserId'
          example: 12
        - name: user_session_<user_id>
          in: cookie
          required: true
          schema:
            type: string
            example: "gAWVQAAAAAAAAACMPCQyYiQxMiRNZE5uT2s2bEROSjFRd2JWeXp5dDN1NkpReTg1eDRpLlVHVUJiTGpoZTM2WkxJZGhscXloT5Qu"
        - name: is_superuser
          in: cookie
          required: true
          schema:
            type: boolean
            example: false
      responses:
        200:
          description: |-
            When an all very well
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileProperties/oneOf/0"
              example:
                id: 30
                original_name: "README.docx"
                size: 27981
                upload_date: "2025-01-18T17:18:03.154301+07:00"
                last_downloaded: null
                comment: null
                file_path: "/uploads/2025/1/18/README.docx"
                special_link: "dda8958e-1711-464e-9d47-acd292241149"
        400:
          description: |-
            When an all very well
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorFile"
              example:
                error: "'pk' is invalid"
  /files/{pk}/update_comment/:
    post:
        tags:
          - file
        summary: This for change (rewrite) a comment to a single file
        operationId: FileStorageViewSet
        requestBody:
            required: true
            content:
              application/json:
                schema:
                    type: object
                    properties:
                      comment:
                        type: string
                        example: "my first comment"

        parameters:
            - name: pk
              description: |-
                This is an individual nuber of single file from db.
              in: path
              required: true
              schema:
                type: integer
                example: 32
            - name: user_session_<user_id>
              description: |-
                This. the user receives from registration.
              in: cookie
              required: true
              schema:
                type: string
                example: "gAWVQAAAAAAAAACMPCQyYiQxMiRNZE5uT2s2bEROSjFRd2JWeXp5dDN1NkpReTg1eDRpLlVHVUJiTGpoZTM2WkxJZGhscXloT5Qu"
            - name: is_superuser_<user_id>
              description: |-
                This. the user receives from registration.
              in: cookie
              required: true
              schema:
                type: boolean
                example: false
        responses:
            200:
                description: |-
                    The comment was changed.
                headers:
                    Set-Cookie:
                        schema:
                            allOf:
                                - type: string
                                  example: user_session_<user_id>=gAWVQAAAAAAAAACMPCQyYiQxMiRNZE5uT2s2bEROSjFRd2JWeXp5dDN1NkpReTg1eDRpLlVHVUJiTGpoZTM2WkxJZGhscXloT5Qu; Path=/; HttpOnly
                                  description: Cookie for user session
                                - type: string
                                  example: is_superuser_<user_id>=false; Path=/; HttpOnly
                                  description: Cookie for user session
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/FileProperties/oneOf/0'
                        example:
                            id: 30
                            original_name: "README.docx"
                            size: 27981
                            upload_date: "2025-01-18T17:18:03.154301+07:00"
                            last_downloaded: null
                            comment: "my first comment"
                            file_path: "/uploads/2025/1/18/README.docx"
                            special_link: "dda8958e-1711-464e-9d47-acd292241149"
            403:
                description: Permission denied
                headers:
                  Set-Cookie:
                    schema:
                        allOf:
                            - type: string
                              example: user_session_<user_id>=gAWVQAAAAAAAAACMPCQyYiQxMiRNZE5uT2s2bEROSjFRd2JWeXp5dDN1NkpReTg1eDRpLlVHVUJiTGpoZTM2WkxJZGhscXloT5Qu; Path=/; HttpOnly
                              description: Cookie for user session
                            - type: string
                              example: is_superuser_<user_id>=false; Path=/; HttpOnly
                              description: Cookie for user session
                content:
                  application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorFile'
                    example:
                      error: "Permission denied."
            404:
                description: File not found.
                headers:
                  Set-Cookie:
                    schema:
                        allOf:
                          - type: string
                            example: user_session_<user_id>=gAWVQAAAAAAAAACMPCQyYiQxMiRNZE5uT2s2bEROSjFRd2JWeXp5dDN1NkpReTg1eDRpLlVHVUJiTGpoZTM2WkxJZGhscXloT5Qu; Path=/; HttpOnly
                            description: Cookie for user session
                          - type: string
                            example: is_superuser_<user_id>=false; Path=/; HttpOnly
                            description: Cookie for user session
                content:
                    application/json:
                      schema:
                        $ref: '#/components/schemas/ErrorFile'
                      example:
                        error: "File not found."
...
components:
  schemas:
    UserId:
      type: integer
      properties:
        pk:
          type: integer
          nullable: false
          minimum: 0
          example: 1
...
    ErrorFile:
      type: string
      required:
          - error
      properties:
          error:
            type: string
            example:  No file provided

#
# securitySchemes:
#     ApiKey:
#       type: apiKey
#       in: header
#       name: X-Api-Key



# security:
#   - ApiKey: []